\name{sdat-package}
\alias{sdat-package}
\alias{sdat}
\docType{package}
\title{
\packageTitle{sdat}
}
\description{
\packageDescription{sdat}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{sdat}
\packageIndices{sdat}
The main function is \code{marginal.test}.
}
\author{
\packageAuthor{sdat}

Maintainer: \packageMaintainer{sdat}
}
